requirements.txt
streamlit>=1.28.0
mistralai>=0.1.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.17.0
fpdf>=1.7.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
python-dateutil>=2.8.0